requirements.txt
ipython
ipdb
django==1.6.3
djangorestframework==2.3.13
gunicorn
celery
south
psycopg2
redis
hiredis
jsonfield
pystache
pillow
pytz
tablib
factory-boy
coverage
sphinx
raven
ckanclient
unicode-slugify
django-oauth-toolkit
django-debug-toolbar
django-cors-headers
django-redis-cache
django-autoslug
django-gravatar2
django-taggit
django-braces
django-subdomains
django-uuidfield

#django-cacheops: master can serialize uuid fields
git+https://github.com/Suor/django-cacheops.git

# upstream still doesn't fully support custom user model
hg+https://bitbucket.org/prjts/django-registration

#quilt==0.2.0-alpha
git+https://github.com/pwalsh/quilt#egg=quilt

#dock==0.2.0-alpha
git+https://github.com/pwalsh/dock#egg=dock
